# homework06/hasker/Hasker-1.0.0-swagger.yaml
swagger: '2.0'
info:
  description: Пример API для проекта "Hasker" на курсе OTUS
  version: 0.1.0
  title: Hasker API
  contact:
    email: dementiy@yandex.ru
consumes:
- application/json
produces:
- application/json
paths:
...
  /trending:
    get:
      summary: Получение списка популярных вопросов
      operationId: trending
      description: |
        Список популярных вопросов с пагинацией по 5 вопросов на страницу. 
        Вопросы упорядочены по числу голосов и дате создания.
      parameters:
      - in: query
        name: page
        description: Номер страницы.
        required: false
        type: integer
      responses:
        200:
          description: Упорядоченный список вопросов.
          schema:
            type: array
            items:
              $ref: '#/definitions/Questions'
        404:
          description: Не верно указан номер страницы для пагинации.
...
definitions:
  Question:
    description: |
      Вопрос.
    type: object
    properties:
      id:
        type: number
        format: int64
        description: Идентификатор данного вопроса.
        readOnly: true
      title:
        type: string
        format: text
        description: Заголовок вопроса.
        example: Python или Go?
        x-isnullable: false
      text:
        type: string
        format: text
        description: Текст вопроса.
        example: Что лучше Python или Go?
        x-isnullable: false
      author:
        type: string
        format: identity
        description: Автор, задавший данный вопрос.
        example: dementiy
        x-isnullable: false
      answered:
        type: boolean
        description: |
          Истина, если вопрос помечен автором как "получен подходящий ответ".
        readOnly: true
        x-isnullable: false
      total_votes:
        type: number
        format: int64
        description: Число голосов за данный вопрос.
        readOnly: true
      tags:
        $ref: '#/definitions/Tags'
      created:
        type: string
        format: date-time
        description: Дата создания вопроса.
        readOnly: true
        x-isnullable: true
  Questions:
    type: array
    items:
      $ref: '#/definitions/Question'
  Tag:
    description: Тег
    type: object
    properties:
      name:
        type: string
        format: text
        description: Имя тега.
        readOnly: true
  Tags:
    type: array
    items:
      $ref: '#/definitions/Tag'
...
basePath: /api
schemes:
 - http